Refactor operator main
Refactor operator main and move probes server into separate file to keep
the main file lean. This is in preparation for adding leader election
code which will require a bit more rework.
