+++ cmd/transflect-operator/main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
@@ -17,7 +16,10 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
-var version = "v0.0.0"
+var (
+	version   = "v0.0.0"
+	errSignal = fmt.Errorf("received shutdown signal")
+)
 
 type config struct {
 	UseIngress bool             `short:"i" help:"Create and use temporary ingress to access temporary service, e.g. from outside cluster"`
@@ -36,40 +38,6 @@ func main() {
 		os.Exit(1)
 	}
 }
-
-type probesServer struct {
-	http.Server
-}
-
-func newProbesServer() *probesServer {
-	ok := func(w http.ResponseWriter, r *http.Request) {
-		fmt.Fprintf(w, "OK")
-	}
-	mux := http.NewServeMux()
-	mux.HandleFunc("/_liveness", ok)
-	mux.HandleFunc("/_readiness", ok)
-	return &probesServer{
-		Server: http.Server{
-			Addr:    ":8080",
-			Handler: mux,
-		},
-	}
-}
-
-func (s *probesServer) start() error {
-	if err := s.ListenAndServe(); !errors.Is(err, http.ErrServerClosed) {
-		return errors.Wrap(err, "cannot start probes server")
-	}
-	return nil
-}
-
-func (s *probesServer) stop() {
-	if err := s.Shutdown(context.Background()); err != nil {
-		log.Logger.Error().Err(err).Msg("cannot stop HTTP server")
-	}
-}
-
-var errSignal = fmt.Errorf("received shutdown signal")
 
 func run(cfg *config) error {
 	setupLogging(cfg)
